Allows directories to be linked to.
This is basically a bugfix whereby directories
were turned into a .html link. That is, if you
had a link like this [dir](dir/) it would
ultimately turn into <a href="dir.html">dir</a>

This fixes it so that directories can be linked to.
Basically we ensure that the filename _is_ a markdown
file, and in that case we do alter the linked file
to be .html extension. Any other files are ignored.