Use `CGI` instead of `EscapeUtils`
Since Ruby 2.5, the escape methods in `CGI` were optimized to the point
it makes no sense to use `EscapeUtils`. Considering Ruby 2.5, and even
2.6, are EOL[1] I believe it's fine to simply change the call to `CGI`.

This commit changes `EscapeUtils.escape_html` to `CGI.escape_html`.

Since the `html-pipeline` gem requires `escape_utils`, it isn't possible
to remove the dependency from the gemspec. I might open an issue or PR
there asking if it's possible to accomodate not requiring it in newer
Ruby versions.

[1]: https://www.ruby-lang.org/en/downloads/branches/